fixes clan page error
