+++ components/thisUserProfile/profileComponent.js
@@ -8,13 +8,20 @@ import { fetchSingleProfile } from '../../store/features/singleProfile'
 export default function ProfilePage() {
   const dispatch = useDispatch()
   useEffect(() => {
-    dispatch(fetchHouseholdInfo())
     dispatch(fetchSingleProfile())
     return () => {
-      dispatch(fetchHouseholdInfo())
       dispatch(fetchSingleProfile())
     }
   }, [])
+
+  useEffect(() => {
+    if (profile && profile.id) {
+      dispatch(fetchHouseholdInfo())
+    }
+    return () => {
+      dispatch(fetchHouseholdInfo())
+    }
+  }, [profile])
 
   let { singleProfile, singleHouseholdProfiles } = useSelector((store) => store)
   let [profile, profileLoading] = [singleProfile.profile, singleProfile.loading]
@@ -38,28 +45,6 @@ export default function ProfilePage() {
         ></div>
       </div>
     )
-  const easing = [0.6, -0.05, 0.01, 0.99]
-  const stagger = {
-    animate: {
-      transition: {
-        staggerChildren: 0.2,
-      },
-    },
-  }
-  const fadeInUp = {
-    initial: {
-      y: 50,
-      opacity: 0,
-    },
-    animate: {
-      y: 0,
-      opacity: 1,
-      transition: {
-        duration: 0.6,
-        ease: easing,
-      },
-    },
-  }
   //after loading, display content components
   if (!profileLoading && !householdLoading)
     return (
