+++ store/features/houseProfiles.js
@@ -7,28 +7,29 @@ const initialState = {
   loading: false,
 }
 
-const user = supabase.auth.user()
-
 // *** THUNKS *** //
 // fetch all profiles from a household
 export const fetchAllProfiles = createAsyncThunk(
   'household/fetchAllProfiles',
   async () => {
-    try {
-      let { data: houseHoldId } = await supabase
-        .from('profiles')
-        .select('household_id')
-        .eq('id', user.id)
-        .single()
+    const user = supabase.auth.user()
+    if (user && user.id) {
+      try {
+        let { data: houseHoldId } = await supabase
+          .from('profiles')
+          .select('household_id')
+          .eq('id', user.id)
+          .single()
 
-      let { data: profiles } = await supabase
-        .from('profiles')
-        .select('*')
-        .eq('household_id', houseHoldId.household_id)
-      return profiles
-    } catch (error) {
-      console.log(error)
-      return error
+        let { data: profiles } = await supabase
+          .from('profiles')
+          .select('*')
+          .eq('household_id', houseHoldId.household_id)
+        return profiles
+      } catch (error) {
+        console.log(error)
+        return error
+      }
     }
   }
 )
@@ -37,7 +38,8 @@ export const fetchAllProfiles = createAsyncThunk(
 export const fetchHouseholdInfo = createAsyncThunk(
   'household/householdInfo',
   async () => {
-    if (user.id) {
+    const user = supabase.auth.user()
+    if (user && user.id) {
       try {
         let { data: userProfile } = await supabase
           .from('profiles')
